chore: migrate classes to common (#409)
