+++ CustomerIOMessagingInApp.podspec
@@ -20,5 +20,5 @@ Pod::Spec.new do |spec|
   spec.source_files  = "Sources/MessagingInApp/**/*"
   spec.module_name = "CioMessagingInApp"  # the `import X` name when using SDK in Swift files
   
-  spec.dependency "CustomerIOTracking", "= #{spec.version.to_s}"
+  spec.dependency "CustomerIOCommon", "= #{spec.version.to_s}"
  end
