+++ Makefile
@@ -7,10 +7,10 @@ SHELL = /bin/sh
 generate:
 	./binny sourcery --sources Sources/Common --templates Sources/Templates --output Sources/Common/autogenerated 
 	./binny sourcery --sources Sources/Tracking --templates Sources/Templates --output Sources/Tracking/autogenerated --args imports=CioInternalCommon
-	./binny sourcery --sources Sources/MessagingPush --templates Sources/Templates --output Sources/MessagingPush/autogenerated --args imports=CioTracking-CioInternalCommon
-	./binny sourcery --sources Sources/MessagingPushAPN --templates Sources/Templates --output Sources/MessagingPushAPN/autogenerated --args imports=CioMessagingPush-CioTracking-CioInternalCommon
-	./binny sourcery --sources Sources/MessagingPushFCM --templates Sources/Templates --output Sources/MessagingPushFCM/autogenerated --args imports=CioMessagingPush-CioTracking-CioInternalCommon
-	./binny sourcery --sources Sources/MessagingInApp --templates Sources/Templates --output Sources/MessagingInApp/autogenerated --args imports=CioTracking-CioInternalCommon
+	./binny sourcery --sources Sources/MessagingPush --templates Sources/Templates --output Sources/MessagingPush/autogenerated --args imports=CioInternalCommon
+	./binny sourcery --sources Sources/MessagingPushAPN --templates Sources/Templates --output Sources/MessagingPushAPN/autogenerated --args imports=CioMessagingPush-CioInternalCommon
+	./binny sourcery --sources Sources/MessagingPushFCM --templates Sources/Templates --output Sources/MessagingPushFCM/autogenerated --args imports=CioMessagingPush-CioInternalCommon
+	./binny sourcery --sources Sources/MessagingInApp --templates Sources/Templates --output Sources/MessagingInApp/autogenerated --args imports=CioInternalCommon
 
 lint:
 	./binny swiftlint lint --strict
