+++ Sources/Common/Util/CleanupRepository.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol CleanupRepository: AutoMockable {
+    func cleanup()
+}
