+++ Sources/Common/Util/DeviceMetricsGrabber.swift
@@ -5,6 +5,7 @@ protocol DeviceMetricsGrabber: AutoMockable {
 }
 
 // sourcery: InjectRegister = "DeviceMetricsGrabber"
+// sourcery: InjectRegisterShared = "DeviceMetricsGrabber"
 class DeviceMetricsGrabberImpl: DeviceMetricsGrabber {
     var appBundleId: String? {
         Bundle.main.bundleIdentifier
