+++ Sources/Common/Util/Log.swift
@@ -111,6 +111,47 @@ public class ConsoleLogger: Logger {
     }
 }
 
+// sourcery: InjectRegisterShared = "Logger"
+public class SharedConsoleLogger: Logger {
+    // allows filtering in Console mac app
+    public static let logSubsystem = "io.customer.sdk"
+    public static let logCategory = "CIO"
+
+    private func printMessage(_ message: String, _ level: CioLogLevel) {
+        ConsoleLogger.logMessageToConsole(message, level: level)
+    }
+
+    public func debug(_ message: String) {
+        printMessage(message, .debug)
+    }
+
+    public func info(_ message: String) {
+        printMessage("ℹ️ \(message)", .info)
+    }
+
+    public func error(_ message: String) {
+        printMessage("🛑 \(message)", .error)
+    }
+
+    public static func logMessageToConsole(_ message: String, level: CioLogLevel) {
+        #if canImport(os)
+        // Unified logging for Swift. https://www.avanderlee.com/workflow/oslog-unified-logging/
+        // This means we can view logs in xcode console + Console app.
+        if #available(iOS 14, *) {
+            let logger = os.Logger(subsystem: self.logSubsystem, category: self.logCategory)
+            logger.log(level: level.osLogLevel, "\(message, privacy: .public)")
+        } else {
+            let logger = OSLog(subsystem: logSubsystem, category: logCategory)
+            os_log("%{public}@", log: logger, type: level.osLogLevel, message)
+        }
+        #else
+        // At this time, Linux cannot use `os.log` or `OSLog`. Instead, use: https://github.com/apple/swift-log/
+        // As we don't officially support Linux at this time, no need to add a dependency to the project.
+        // therefore, we are not logging if can't import os.log
+        #endif
+    }
+}
+
 // Log messages to customers when the SDK is not initialized.
 // Great for alerts when the SDK may not be setup correctly.
 // Since the SDK is not initialized, dependencies graph is not created.
