+++ Sources/Common/Util/ThreadUtil.swift
@@ -7,6 +7,7 @@ public protocol ThreadUtil {
 }
 
 // sourcery: InjectRegister = "ThreadUtil"
+// sourcery: InjectRegisterShared = "ThreadUtil"
 public class CioThreadUtil: ThreadUtil {
     public func runMain(_ block: @escaping () -> Void) {
         DispatchQueue.main.async(execute: block)
