+++ Sources/MessagingPushFCM/MessagingPushFCM.swift
@@ -1,6 +1,5 @@
 import CioInternalCommon
 import CioMessagingPush
-import CioTracking
 import Foundation
 #if canImport(UserNotifications)
 import UserNotifications
