+++ Sources/Templates/AutoDependencyInjection.stencil
@@ -118,4 +118,40 @@ extension DIGraph {
     {% endfor %} 
 }
 
+extension DIGraphShared {
+    // Handle classes annotated with InjectRegisterShared
+    {% for dep in types.all where dep|annotated:"InjectRegisterShared" %}
+    {% set class %}{{ dep.annotations["InjectRegisterShared"] }}{% endset %}
+    
+    {% if dep|annotated:"InjectCustomShared" %}
+    // {{ class }} (custom. property getter provided via extension)
+    {% call overrideGetterProperty class dep true false %}
+    {% elif dep|annotated:"InjectSingletonShared" %}
+    // {{ class }} (singleton)
+    {% call overrideGetterProperty class dep false true %}
+    {{ dep.accessLevel }} var shared{{ class }}: {{ class }} {
+        return DispatchQueue(label: "DIGraphShared_{{class}}_singleton_access").sync {
+            if let overridenDep: {{ class }} = getOverriddenInstance() {
+                return overridenDep
+            }
+            let existingSingletonInstance = self.singletons[String(describing: {{ class }}.self)] as? {{ class }}
+            let instance = existingSingletonInstance ?? _get_{{ class|lowerFirstLetter }}()
+            self.singletons[String(describing: {{ class }}.self)] = instance
+            return instance
+        }
+    }
+    private func _get_{{ class|lowerFirstLetter }}() -> {{ class }} {
+        return {{ dep.name }}({% for initializer in dep.initializers %}{% if not initializer.isConvenienceInitializer %}{% for param in initializer.parameters %}{{ param.name }}: self.{{ param.typeName|lowerFirstLetter }}{% if not forloop.last%}, {% endif %}{% endfor %}{% endif %}{% endfor %})
+    }
+    {% else %}
+    // {{ class }}
+    {% call overrideGetterProperty class dep false false %}
+    {% call addPropertyWrappers dep %}
+    private var new{{ class }}: {{ class }} {
+        return {{ dep.name}}({% for initializer in dep.initializers %}{% if not initializer.isConvenienceInitializer %}{% for param in initializer.parameters %}{{ param.name }}: self.{{ param.typeName|lowerFirstLetter }}{% if not forloop.last%}, {% endif %}{% endfor %}{% endif %}{% endfor %})
+    }
+    {% endif %}
+    {% endfor %}
+}
+
 // swiftlint:enable all
