+++ Sources/Tracking/Repository/CleanupRepository.swift
@@ -1,10 +1,6 @@
 import CioInternalCommon
 import Foundation
 
-protocol CleanupRepository: AutoMockable {
-    func cleanup()
-}
-
 // sourcery: InjectRegister = "CleanupRepository"
 class CioCleanupRepository: CleanupRepository {
     private let queue: Queue
