+++ Sources/Tracking/Type/Aliases.swift
@@ -13,3 +13,4 @@ public typealias CioLogLevel = CioInternalCommon.CioLogLevel
 public typealias CioSdkConfig = CioInternalCommon.SdkConfig
 public typealias CioNotificationServiceExtensionSdkConfig = CioInternalCommon.NotificationServiceExtensionSdkConfig
 public typealias Metric = CioInternalCommon.Metric
+public typealias CustomerIO = CioInternalCommon.CustomerIO
