+++ Sources/Tracking/autogenerated/AutoDependencyInjection.generated.swift
@@ -84,4 +84,8 @@ extension DIGraph {
     }
 }
 
+extension DIGraphShared {
+    // Handle classes annotated with InjectRegisterShared
+}
+
 // swiftlint:enable all
