+++ Tests/Tracking/CustomerIOMockTest.swift
@@ -1,3 +1,4 @@
+@testable import CioInternalCommon
 @testable import CioTracking
 import Foundation
 import SharedTests
